refactor: update import paths for skeleton components and streamline error handling in various services
